Transferred from the Adobe Open Source Subversion Repository
Revision: 11983
Author:   clucier@adobe.com
Date:     2009-11-19T12:25:01.335470Z

Log Message: 
-----------
Mirroring change we made in 3.x revision 11672 so that we can close down (2) bugs.

QE notes: None 
Doc notes: None
Bugs: SDK-23661, SDK-24274
Reviewer: Alex
Tests run: Checkin
Is noteworthy for integration: No

Modified Paths:
--------------
/flex/sdk/trunk/frameworks/projects/framework/src/mx/controls/ComboBox.as

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk@1268043 13f79535-47bb-0310-9956-ffa450edef68

+++ frameworks/projects/mx/src/mx/controls/ComboBox.as
@@ -1447,7 +1447,11 @@ public class ComboBox extends ComboBase
         // and will not affect the dropdown size
         if (_dropdown && !inTween)
         {
-            destroyDropdown();
+            // We will not destroy our dropdown if user is actively
+            // interacting with it, this has been legacy behavior since
+            // Flex 3.0.
+            if (!_showingDropdown)
+                destroyDropdown();
         }
         else if (_showingDropdown)
         {
